+++ components/schemas/PasswordValidationSchema.js
@@ -1,10 +1,6 @@
 import * as yup from "yup";
 
 export const PasswordValidationSchema = yup.object().shape({
-  email: yup
-    .string()
-    .email("Please enter valid email")
-    .required("Email address is required"),
   password: yup
     .string()
     .min(8, ({ min }) => `Password must be at least ${min} characters`)
